Initial creation of birdseye visual generator
Began creation of birdseye visual generator for training data generation for deep-learning object detection. Unlike the initial plan to randomise position, materials, lighting and backdrop, this version will use existing materials, appearances and identical (or similar) positioning. The result will be a an idealised case, which will closely reflect the simulation environment. While this form of bias and overfitting is generally ill-advised, it will be acceptable for the purpose of benchmarking compute performance rather than absolute accuracy when compared to techniques that do not benefit from this generous data-driven advantage.